Fixed more tests

+++ vmaas_sync/metrics_cyndi_test.go
@@ -27,7 +27,7 @@ func TestUploadedCyndiSystemsCounts1D(t *testing.T) {
 	systemsQuery = updateCyndiQueryLastUpload(systemsQuery, refTime(), 1)
 	var nSystems int64
 	assert.Nil(t, systemsQuery.Count(&nSystems).Error)
-	assert.Equal(t, 16, nSystems)
+	assert.Equal(t, int64(16), nSystems)
 }
 
 func TestUploadedCyndiSystemsCounts7D(t *testing.T) {
@@ -38,7 +38,7 @@ func TestUploadedCyndiSystemsCounts7D(t *testing.T) {
 	systemsQuery = updateCyndiQueryLastUpload(systemsQuery, refTime(), 7)
 	var nSystems int64
 	assert.Nil(t, systemsQuery.Count(&nSystems).Error)
-	assert.Equal(t, 16, nSystems)
+	assert.Equal(t, int64(16), nSystems)
 }
 
 func TestUploadedCyndiSystemsCounts30D(t *testing.T) {
@@ -49,7 +49,7 @@ func TestUploadedCyndiSystemsCounts30D(t *testing.T) {
 	systemsQuery = updateCyndiQueryLastUpload(systemsQuery, refTime(), 30)
 	var nSystems int64
 	assert.Nil(t, systemsQuery.Count(&nSystems).Error)
-	assert.Equal(t, 16, nSystems)
+	assert.Equal(t, int64(16), nSystems)
 }
 
 func TestUploadedCyndiSystemsCountsNoSystems(t *testing.T) {
@@ -61,7 +61,7 @@ func TestUploadedCyndiSystemsCountsNoSystems(t *testing.T) {
 	systemsQuery = updateCyndiQueryLastUpload(systemsQuery, refTime, 30)
 	var nSystems int64
 	assert.Nil(t, systemsQuery.Count(&nSystems).Error)
-	assert.Equal(t, 0, nSystems)
+	assert.Equal(t, int64(0), nSystems)
 }
 
 func TestUploadedCyndiSystemsCountsAllSystems(t *testing.T) {
@@ -72,7 +72,7 @@ func TestUploadedCyndiSystemsCountsAllSystems(t *testing.T) {
 	systemsQuery = updateCyndiQueryLastUpload(systemsQuery, refTime(), -1)
 	var nSystems int64
 	assert.Nil(t, systemsQuery.Count(&nSystems).Error)
-	assert.Equal(t, 16, nSystems)
+	assert.Equal(t, int64(16), nSystems)
 }
 
 func TestCyndiTags(t *testing.T) {
@@ -81,8 +81,8 @@ func TestCyndiTags(t *testing.T) {
 
 	cyndiData, err := getCyndiData()
 	assert.Nil(t, err)
-	assert.Equal(t, 16, cyndiData.SystemsCount)
-	assert.Equal(t, 3, cyndiData.UniqueTags)
-	assert.Equal(t, 14, cyndiData.SapCount)
-	assert.Equal(t, 15, cyndiData.SystemsWithTags)
+	assert.Equal(t, int64(16), cyndiData.SystemsCount)
+	assert.Equal(t, int64(3), cyndiData.UniqueTags)
+	assert.Equal(t, int64(14), cyndiData.SapCount)
+	assert.Equal(t, int64(15), cyndiData.SystemsWithTags)
 }
