+++ vmaas_sync/metrics_test.go
@@ -34,8 +34,8 @@ func TestSystemsCountsOptOut(t *testing.T) {
 	var notOptOuted int64
 	assert.Nil(t, updateSystemsQueryOptOut(systemsQuery, true).Count(&optOuted).Error)
 	assert.Nil(t, updateSystemsQueryOptOut(systemsQuery, false).Count(&notOptOuted).Error)
-	assert.Equal(t, 2, optOuted)
-	assert.Equal(t, 12, notOptOuted)
+	assert.Equal(t, int64(2), optOuted)
+	assert.Equal(t, int64(12), notOptOuted)
 }
 
 func TestUploadedSystemsCounts1D(t *testing.T) {
@@ -46,7 +46,7 @@ func TestUploadedSystemsCounts1D(t *testing.T) {
 	systemsQuery = updateSystemsQueryLastUpload(systemsQuery, refTime(), 1)
 	var nSystems int64
 	assert.Nil(t, systemsQuery.Count(&nSystems).Error)
-	assert.Equal(t, 2, nSystems)
+	assert.Equal(t, int64(2), nSystems)
 }
 
 func TestUploadedSystemsCounts7D(t *testing.T) {
@@ -57,7 +57,7 @@ func TestUploadedSystemsCounts7D(t *testing.T) {
 	systemsQuery = updateSystemsQueryLastUpload(systemsQuery, refTime(), 7)
 	var nSystems int64
 	assert.Nil(t, systemsQuery.Count(&nSystems).Error)
-	assert.Equal(t, 5, nSystems)
+	assert.Equal(t, int64(5), nSystems)
 }
 
 func TestUploadedSystemsCounts30D(t *testing.T) {
@@ -68,7 +68,7 @@ func TestUploadedSystemsCounts30D(t *testing.T) {
 	systemsQuery = updateSystemsQueryLastUpload(systemsQuery, refTime(), 30)
 	var nSystems int64
 	assert.Nil(t, systemsQuery.Count(&nSystems).Error)
-	assert.Equal(t, 8, nSystems)
+	assert.Equal(t, int64(8), nSystems)
 }
 
 func TestUploadedSystemsCountsNoSystems(t *testing.T) {
@@ -80,7 +80,7 @@ func TestUploadedSystemsCountsNoSystems(t *testing.T) {
 	systemsQuery = updateSystemsQueryLastUpload(systemsQuery, refTime, 30)
 	var nSystems int64
 	assert.Nil(t, systemsQuery.Count(&nSystems).Error)
-	assert.Equal(t, 1, nSystems)
+	assert.Equal(t, int64(1), nSystems)
 }
 
 func TestUploadedSystemsCountsAllSystems(t *testing.T) {
@@ -91,7 +91,7 @@ func TestUploadedSystemsCountsAllSystems(t *testing.T) {
 	systemsQuery = updateSystemsQueryLastUpload(systemsQuery, refTime(), -1)
 	var nSystems int64
 	assert.Nil(t, systemsQuery.Count(&nSystems).Error)
-	assert.Equal(t, 14, nSystems)
+	assert.Equal(t, int64(14), nSystems)
 }
 
 func TestAdvisoryCounts(t *testing.T) {
@@ -100,10 +100,10 @@ func TestAdvisoryCounts(t *testing.T) {
 
 	unknown, enh, bug, sec, err := getAdvisoryCounts()
 	assert.Nil(t, err)
-	assert.Equal(t, 0, unknown)
-	assert.Equal(t, 3, enh)
-	assert.Equal(t, 3, bug)
-	assert.Equal(t, 3, sec)
+	assert.Equal(t, int64(0), unknown)
+	assert.Equal(t, int64(3), enh)
+	assert.Equal(t, int64(3), bug)
+	assert.Equal(t, int64(3), sec)
 }
 
 func TestSystemAdvisoriesStats(t *testing.T) {
